refactoring splunk
