+++ stix_shifter_modules/splunk/stix_translation/Usage.md
