+++ stix_shifter_modules/splunk/stix_translation/json/to_stix_map.json
