+++ stix_shifter_modules/splunk/stix_translation/object_scopers.py
