+++ stix_shifter_modules/splunk/stix_translation/query_constructor.py
