+++ stix_shifter_modules/splunk/stix_translation/splunk_translator.py
@@ -1,6 +1,5 @@
-# ..base.base_translator import BaseTranslator
-from ..base.base_translator import BaseTranslator
-from ...json_to_stix.json_to_stix import JSONToStix
+from stix_shifter_utils.modules.base.stix_translation.base_translator import BaseTranslator
+from stix_shifter_utils.stix_translation.src.json_to_stix.json_to_stix import JSONToStix
 from .stix_to_query import StixToQuery
 from .splunk_utils import hash_type_lookup
