+++ stix_shifter_modules/splunk/stix_translation/splunk_utils.py
