+++ stix_shifter_modules/splunk/stix_transmission/splunk_auth.py
