+++ stix_shifter_modules/splunk/stix_transmission/splunk_ping.py
@@ -1,6 +1,6 @@
-from ..base.base_ping import BasePing
+from stix_shifter_utils.modules.base.stix_transmission.base_ping import BasePing
 import json
-from .....utils.error_response import ErrorResponder
+from stix_shifter_utils.utils.error_response import ErrorResponder
 
 class SplunkPing(BasePing):
     def __init__(self, api_client):
