+++ stix_shifter_modules/splunk/stix_transmission/splunk_query_connector.py
@@ -1,6 +1,6 @@
-from ..base.base_query_connector import BaseQueryConnector
+from stix_shifter_utils.modules.base.stix_transmission.base_connector import BaseQueryConnector
 import json
-from .....utils.error_response import ErrorResponder
+from stix_shifter_utils.utils.error_response import ErrorResponder
 
 
 class SplunkQueryConnector(BaseQueryConnector):
