+++ stix_shifter_modules/splunk/stix_transmission/splunk_results_connector.py
@@ -1,7 +1,7 @@
-from ..base.base_results_connector import BaseResultsConnector
+from stix_shifter_utils.modules.base.stix_transmission.base_results_connector import BaseResultsConnector
 from .spl_api_client import APIClient
 import json
-from .....utils.error_response import ErrorResponder
+from stix_shifter_utils.utils.error_response import ErrorResponder
 
 
 class SplunkResultsConnector(BaseResultsConnector):
