+++ stix_shifter_utils/utils/error_response.py
@@ -107,9 +107,10 @@ class ErrorResponder():
     @staticmethod
     def call_module_error_mapper(json_data, return_object):
         caller_path_list = traceback.extract_stack()[-3].filename.split('/')
-        caller_module_name = caller_path_list[-2]
-        path_start_position = ErrorResponder.rindex(caller_path_list, 'stix_shifter')
-        module_path = '.'.join(caller_path_list[path_start_position: -1]) + '.' + caller_module_name + '_error_mapper'
+        caller_module_name = caller_path_list[-3:-1]
+        # path_start_position = ErrorResponder.rindex(caller_path_list, 'stix-shifter')
+        module_path = 'stix_shifter_modules.' + caller_module_name[0] + '.'  + caller_module_name[1] + '.' + caller_module_name[0] + '_error_mapper'
+        # module_path = '.'.join(caller_path_list[path_start_position: -1]) + '.' + caller_module_name + '_error_mapper'
         try:
             module = importlib.import_module(module_path)
             if json_data is not None:
